Geolocation moved to the contacts fragment.
